Potentiometer values in arduino are now uint16_t instead of an array of both pin values
Because of the change some changes were needed in the python files

+++ ArduinoFiles/BitBang/Components/Button/Button.cpp
@@ -9,7 +9,7 @@ Button::Button(int pin, uint8_t id): Component(id, 'B') {
 }
 
 //Get the value of the pin the button is connected to
-uint8_t Button::getValue() {
+uint16_t Button::getValue() {
     uint8_t value = digitalRead(_pin);
     _hasChanged = (_prevValue != value);
     _prevValue = value;
